refactor: reorganize the library arch.
This is a decision that the core development team met on. The goal of this
refactor is to provide a better and more organized library architecture
that is easier to navigate around for us and collaborators wanting to get
involved.
