+++ interactions/api/enums.py
@@ -1,4 +1,3 @@
-# Normal libraries
 from enum import IntEnum
 
 # TODO: post-v4: Implement this into the new error system at a later point.
