+++ interactions/api/gateway/__init__.py
@@ -1,0 +1,8 @@
+"""
+interactions.api.gateway
+
+This section of the library maintains and
+handles all of the Gateway work.
+"""
+from .client import *  # noqa: F401 F403
+from .heartbeat import *  # noqa: F401 F403
