+++ interactions/api/gateway/client.pyi
@@ -8,20 +8,16 @@ from typing import Any, Dict, List, Optional, Tuple, Union, Iterable
 
 from aiohttp import ClientWebSocketResponse
 
-from ..models import Option
-from ..api.models.misc import MISSING
-from ..api.models.presence import ClientPresence
-from .dispatch import Listener
-from .http.client import HTTPClient
-from .models.flags import Intents
+from .heartbeat import _Heartbeat
+from ...client.models import Option
+from ...api.models.misc import MISSING
+from ...api.models.presence import ClientPresence
+from ..dispatch import Listener
+from ..http.client import HTTPClient
+from ..models.flags import Intents
 
 log: Logger
 __all__: Iterable[str]
-
-class _Heartbeat:
-    event: Event
-    delay: float
-    def __init__(self, loop: AbstractEventLoop) -> None: ...
 
 class WebSocketClient:
     _loop: AbstractEventLoop
@@ -42,7 +38,6 @@ class WebSocketClient:
     _last_ack: float
     latency: float
     ready: Event
-
     def __init__(
         self,
         token: str,
@@ -53,7 +48,9 @@ class WebSocketClient:
     async def _manage_heartbeat(self) -> None: ...
     async def __restart(self): ...
     async def _establish_connection(
-        self, shard: Optional[List[Tuple[int]]] = MISSING, presence: Optional[ClientPresence] = MISSING
+        self,
+        shard: Optional[List[Tuple[int]]] = MISSING,
+        presence: Optional[ClientPresence] = MISSING,
     ) -> None: ...
     async def _handle_connection(
         self,
@@ -64,7 +61,9 @@ class WebSocketClient:
     async def wait_until_ready(self) -> None: ...
     def _dispatch_event(self, event: str, data: dict) -> None: ...
     def __contextualize(self, data: dict) -> object: ...
-    def __sub_command_context(self, data: Union[dict, Option], _context: Optional[object] = MISSING) -> Union[Tuple[str], dict]: ...
+    def __sub_command_context(
+        self, data: Union[dict, Option], _context: Optional[object] = MISSING
+    ) -> Union[Tuple[str], dict]: ...
     def __option_type_context(self, context: object, type: int) -> dict: ...
     @property
     async def __receive_packet_stream(self) -> Optional[Dict[str, Any]]: ...
