+++ interactions/api/gateway/heartbeat.pyi
@@ -1,0 +1,6 @@
+from asyncio import AbstractEventLoop, Event
+
+class _Heartbeat:
+    event: Event
+    delay: float
+    def __init__(self, loop: AbstractEventLoop) -> None: ...
