+++ interactions/client/context.py
@@ -1,14 +1,14 @@
 from logging import Logger
 from typing import List, Optional, Union
 
-from .api import InteractionException
-from .api.models.channel import Channel
-from .api.models.guild import Guild
-from .api.models.member import Member
-from .api.models.message import Embed, Message, MessageInteraction, MessageReference
-from .api.models.misc import MISSING, DictSerializerMixin, Snowflake
-from .api.models.user import User
-from .base import get_logger
+from ..api import InteractionException
+from ..api.models.channel import Channel
+from ..api.models.guild import Guild
+from ..api.models.member import Member
+from ..api.models.message import Embed, Message, MessageInteraction, MessageReference
+from ..api.models.misc import MISSING, DictSerializerMixin, Snowflake
+from ..api.models.user import User
+from ..base import get_logger
 from .enums import InteractionCallbackType, InteractionType
 from .models.command import Choice
 from .models.component import ActionRow, Button, Modal, SelectMenu, _build_components
