+++ interactions/client/context.pyi
@@ -1,18 +1,17 @@
 from typing import List, Optional, Union
 
-from .api.http.client import HTTPClient
-from .api.models.channel import Channel as Channel
-from .api.models.guild import Guild as Guild
-from .api.models.member import Member as Member
-from .api.models.message import Embed as Embed
-from .api.models.message import Message as Message
-from .api.models.message import MessageInteraction as MessageInteraction
-from .api.models.message import MessageReference as MessageReference
-from .api.models.misc import MISSING as MISSING
-from .api.models.misc import DictSerializerMixin as DictSerializerMixin
-from .api.models.misc import Snowflake as Snowflake
-from .api.models.user import User as User
-from .base import get_logger as get_logger
+from ..api.http.client import HTTPClient
+from ..api.models.channel import Channel as Channel
+from ..api.models.guild import Guild as Guild
+from ..api.models.member import Member as Member
+from ..api.models.message import Embed as Embed
+from ..api.models.message import Message as Message
+from ..api.models.message import MessageInteraction as MessageInteraction
+from ..api.models.message import MessageReference as MessageReference
+from ..api.models.misc import MISSING as MISSING
+from ..api.models.misc import DictSerializerMixin as DictSerializerMixin
+from ..api.models.misc import Snowflake as Snowflake
+from ..api.models.user import User as User
 from .enums import InteractionCallbackType as InteractionCallbackType
 from .enums import InteractionType as InteractionType
 from .models.command import Choice as Choice
