+++ interactions/client/decor.py
@@ -1,7 +1,7 @@
 from typing import Any, Dict, List, Optional, Union
 
-from .api.models.guild import Guild
-from .api.models.misc import MISSING
+from ..api.models.guild import Guild
+from ..api.models.misc import MISSING
 from .enums import ApplicationCommandType
 from .models.command import ApplicationCommand, Option
 from .models.component import Button, Component, SelectMenu
