+++ interactions/client/models/command.py
