+++ interactions/ext/__init__.py
@@ -7,5 +7,6 @@ for making 3rd party library creation
 easy and simplified.
 """
 from .base import *  # noqa: F401 F403
+from .converter import *  # noqa: F401 F403
 from .error import *  # noqa: F401 F403
 from .version import *  # noqa: F401 F403
